"isspace()" expects an "int", not a "char" ...

+++ MonetDB/src/utils/Mx/Code.c
@@ -43,8 +43,8 @@ again:  switch( d->d_dir ){
 	case Bfile:
 	case Ofile:{ char *s;
 	    fname= d->d_cmd;
-	    for(s=fname; *s && !isspace(*s); s++);
-	    if( isspace(*s)) *s =0; 
+	    for(s=fname; *s && !isspace((int)(*s)); s++);
+	    if( isspace((int)(*s))) *s =0; 
 	    if( s==fname){
 			Error("File name missing. %s",d->d_line);
 	    }
